refactor: . pull addGroupHeading() up
